Remove warnign about LOCATION proeprty (SDK-26)
With CMkae 2.8.12 and newer, you can't assume to get the target location at
configure time.  CMake generates the warning:

CMake Warning (dev) at TESTING/CMakeLists.txt:27 (get_target_property):
  Policy CMP0026 is not set: Disallow use of the LOCATION target property.
  Run "cmake --help-policy CMP0026" for policy details.  Use the cmake_policy
  command to set the policy and suppress this warning.

  The LOCATION property should not be read from target "s_test".  Use the
  target name directly with add_custom_command, or use the generator
  expression $<TARGET_FILE>, as appropriate.

But there was no need for this since all of the usages of this were for the
current binary dir.  If the location of the input files needs to be
generalized, then the macro can be enhanced to allow for this.