Remove CACHE from CMAKE_C_FLAGS (SDK-26)
I made a mistake when I created the statement:

  set(CMAKE_C_FLAGS "-DPRNTlevel=0 -DAdd_ ${CMAKE_C_FLAGS}" CACHE STRING "")

With this, if the user set CMAKE_C_FLAGS, it would not prepend DPRNTlevel=0
-DAdd_.  That is *not* what we want.  What we want is to allow the user to add
new C flags and perhaps override the defaults being set here (for example
-DPrintlevel).  By removing the CACHE args, we get the behavior we want.

I also changed the set for CMAKE_C_FLAGS_RELEASE to only set the default and
not prepend the flags.  Given that only -O3 is set, that seems reasonble.