Allow configure of make.inc but remove from VC (SDK-26)
Note that this has the disadvantage that you can't do multiple builds of
SuperLU from the same directory.  Also, if someone does an out-of-soruce build
of SuperLU, then they will modify this file and that may suprise a user trying
to build the example.

Also not that by hard-coding the library with *.a, this can't possibily work
for shared libraries.  This also will not work because the name of the superlu
library is incorrect, for example.