Get Makefile example working again for static builds (SDK-26)
Now this works if you do this (and only this):

$ mkdir build
$ cd build/
$ env CC=gcc cmake ../
$ make
$ cd EXAMPLE/
$ make

This only works for static builds and only if you build under build/ in the
source tree. This would not be too hard to fix but that is something that
someone else can do.  At the very least, the documentation should be updated
to state these restrictions.

I also added a .gitignore file for all of the stuff created by the in-source
build of the raw makefile example.